requirements.txt
# ── Existing ML dependencies ───────────────────────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# ── LangChain & LangGraph ──────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0

# ── Vector store (RAG) ────────────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── MCP ───────────────────────────────────────────────────────────────────────
mcp>=1.0.0

# ── Web / API ──────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Utilities ─────────────────────────────────────────────────────────────────
pyyaml>=6.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
python-docx>=1.1.0
